+++ conf/script/build_system/compiler/host/get_info/version/msvc/__main__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import cli_fetch_compiler_info
-from cli_fetch_msvc_version import cli_fetch_msvc_version
+from cli import cli_fetch_msvc_version
 
 # Run as a script
 if __name__ == '__main__':
